feat: Add interface registration to service collection extension
This commit adds the ability to register interfaces when using the `ServiceRegistrationAttribute` in the `IServiceCollectionExtensions` class. The code now checks if the attribute has the `RegisterInterfaces` property set to true, and if so, it registers all interfaces implemented by the service class. This allows for more flexible and automated dependency injection configuration.