feat: Enable hot reload in MauiAppBuilderExtensions
This commit adds a new method `EnableHotReload` to the `MauiAppBuilderExtensions` class. When called, it sets the `HotReloadService.HotReloadAware` property to true. This change allows for hot reloading of code during debugging.

The `MauiViewManager<TViewModel>` class has also been updated to handle hot reload events. If the `HotReloadService.HotReloadAware` property is true, it subscribes to the `HotReloadService.UpdateApplicationEvent` event and performs necessary actions when a hot reload occurs.

In the sample project, in the `MauiProgram.cs` file, hot reload is enabled using conditional compilation directives (`#if DEBUG`) by calling the newly added method `.EnableHotReload()` on the `MauiAppBuilder`.

Note: The changes related to hot reload are only applicable in debug mode.